♻️ refactor(project): reorganize project structure and imports
- rename and move persona.json from data to config directory for clarity
- move characterDescription to services as characterService for better cohesion
- move replyGenerator to services as replyService for consistent naming
- move memory modules to store directory to reflect persistent cache role
- rename types directory file from types.ts to index.ts for cleaner imports
- update all import paths to reflect new file locations and improve modularity
- add botReady flag in index.ts to prevent message handling before ready
- enhance startup logs to explicitly list loaded commands and bot readiness
- fix regex in fixMathFormatting for correct TeX escaping in Discord messages
- centralize JSON loading with createRequire for consistent persona loading
- maintain consistent import style by adjusting logger imports throughout project
- no functional changes, only structural and import path improvements
