+++ .gitignore
@@ -4,7 +4,7 @@
 .env.local
 package-lock.json
 yarn.lock
-/src/data/persona.json
+/src/config/persona.json
 
 # folders
 build
